Refactorisation majeure pour améliorer la lisibilité, la modularité et la précision du moteur de ray tracing.

- Ajout de bibliothèques `<vector>` et `<filesystem>` dans `RaytracingEngine.cpp`.
- Changement du type `float` à `double` dans `change_luminance` pour une meilleure précision.
- Optimisation des options de compilation dans `RaytracingEngine.vcxproj`.
- Refactorisation des méthodes dans `Scene.h` pour utiliser `std::optional` et simplifier la gestion des intersections.
- Simplification de la génération d'image en supprimant le tuilage et en utilisant OpenMP avec `schedule(dynamic)`.
- Ajout de nouvelles propriétés dans `Material` (`specular`, `refractiveIndex`, etc.) pour enrichir les matériaux.
- Refactorisation de `HitInfo` pour inclure directement les informations de matériau et de normale.
- Amélioration de la gestion des intersections dans `Shape.h` avec des retours `std::optional<HitInfo>`.
- Simplification de la logique de calcul des couleurs et ajout de la méthode `GenerateAntiAliasing`.

Ces changements améliorent les performances, la lisibilité et la modularité du code tout en enrichissant les fonctionnalités de rendu.

+++ RaytracingEngine/RaytracingEngine.cpp
@@ -3,6 +3,7 @@
 #include "Light.h"
 #include "Shape.h"
 #include "Scene.h"
+
 #include <vector>
 #include <filesystem>
 
@@ -49,9 +50,9 @@ double luminance(const Vec3& color)
 	return color.dot(luminanceWeights);
 }
 
-Vec3 change_luminance(Vec3 c_in, float l_out)
+Vec3 change_luminance(Vec3 c_in, double l_out)
 {
-	float l_in = luminance(c_in);
+	double l_in = luminance(c_in);
 	return c_in * (l_out / l_in);
 }
